--- dvc/repo/experiments/executor/__init__.py
@@ -1,4 +1,0 @@
-from .base import BaseExecutor
-from .local import LocalExecutor
-
-__all__ = ["BaseExecutor", "LocalExecutor"]
